Move all errors to errors.py
